+++ support-frontend/assets/pages/contributions-landing/components/StripeCardForm/helpers/dom.ts
@@ -1,0 +1,9 @@
+export function recaptchaElementNotEmpty(): boolean {
+	const el = document.getElementById('robot_checkbox');
+
+	if (el) {
+		return el.children.length > 0;
+	}
+
+	return true;
+}
