+++ support-frontend/assets/pages/contributions-landing/components/StripeCardForm/helpers/errors.ts
@@ -1,0 +1,5 @@
+export function noClientSecretError(json: unknown): Error {
+	return new Error(
+		`Missing client_secret field in server response: ${JSON.stringify(json)}`,
+	);
+}
